Fix invalid mongoose Schema names

+++ src/SlashCommands/Moderation/TempBan.js
@@ -212,7 +212,7 @@ export const SlashCommandF = class extends SlashCommand {
             return;
           }
 
-          await TempBan.deleteOne({ idJoined: `${interaction.guild.id}-${user.user.userid}` });
+          await TempBan.deleteOne({ idJoined: `${interaction.guild.id}-${user.user.userid}` }); //! log user.user.userid just to be sure
 
           const unbanEmbed = new EmbedBuilder()
             .setThumbnail(this.client.user.displayAvatarURL())
